backend/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-split-settings==1.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
